Remove fuzzer v0.
