--- extensions/ipc/fuzz/v0/index.ts
@@ -1,2 +1,0 @@
-export * from "./handler.js";
-export * from "./types.js";
