+++ extensions/ipc/fuzz/v1/handler.test.ts
@@ -9,7 +9,6 @@ import { tinyChainSpec } from "@typeberry/config";
 import { tryAsU8, tryAsU32 } from "@typeberry/numbers";
 import { Result } from "@typeberry/utils";
 import { IpcSender } from "../../server.js";
-import { KeyValue, Version } from "../v0/types.js";
 import { type FuzzMessageHandler, FuzzTarget } from "./handler.js";
 import {
   AncestryItem,
@@ -20,6 +19,8 @@ import {
   MessageType,
   messageCodec,
   PeerInfo,
+  KeyValue,
+  Version,
 } from "./types.js";
 
 const spec = tinyChainSpec;
@@ -28,7 +29,7 @@ class MockV1MessageHandler implements FuzzMessageHandler {
   getPeerInfo: Mock<(value: PeerInfo) => Promise<PeerInfo>> = mock.fn();
   initialize: Mock<(value: Initialize) => Promise<StateRootHash>> = mock.fn();
   importBlock: Mock<(value: BlockView) => Promise<Result<StateRootHash, ErrorMessage>>> = mock.fn();
-  getState: Mock<(value: HeaderHash) => Promise<KeyValue[]>> = mock.fn();
+  getSerializedState: Mock<(value: HeaderHash) => Promise<KeyValue[]>> = mock.fn();
 }
 
 class MockSender extends IpcSender {
@@ -273,7 +274,7 @@ describe("FuzzV1Target Handler", () => {
         value: keyValues,
       };
 
-      mockMessageHandler.getState.mock.mockImplementation(async () => keyValues);
+      mockMessageHandler.getSerializedState.mock.mockImplementation(async () => keyValues);
 
       const fuzzTarget = new FuzzTarget(mockMessageHandler, mockSender, spec);
       await completeHandshake(mockMessageHandler, mockSender, fuzzTarget);
@@ -283,8 +284,8 @@ describe("FuzzV1Target Handler", () => {
 
       await fuzzTarget.onSocketMessage(testMessage);
 
-      assert.strictEqual(mockMessageHandler.getState.mock.callCount(), 1);
-      assert.deepStrictEqual(mockMessageHandler.getState.mock.calls[0].arguments, [headerHash]);
+      assert.strictEqual(mockMessageHandler.getSerializedState.mock.callCount(), 1);
+      assert.deepStrictEqual(mockMessageHandler.getSerializedState.mock.calls[0].arguments, [headerHash]);
 
       assert.strictEqual(mockSender._sentData.length, 1);
       const sentMessage = decodeMessage(mockSender._sentData[0]);
@@ -390,7 +391,7 @@ describe("FuzzV1Target Handler", () => {
       assert.strictEqual(mockMessageHandler.getPeerInfo.mock.callCount(), 0);
       assert.strictEqual(mockMessageHandler.initialize.mock.callCount(), 0);
       assert.strictEqual(mockMessageHandler.importBlock.mock.callCount(), 0);
-      assert.strictEqual(mockMessageHandler.getState.mock.callCount(), 0);
+      assert.strictEqual(mockMessageHandler.getSerializedState.mock.callCount(), 0);
     });
   });
